Implement blob deletion cron.
In order to prevent user mistakes, user-uploaded blobs are not deleted
the moment they are removed from the note. Implement a cron that
periodically parses all of the user-uploaded blob IDs that are present
in all notes, and compares it with IDs of all of the currently stored blobs.
If a blob is present in storage but not referenced in any of the notes,
it is deleted. This way if a user accidentaly deleted a super-important
blob from the note, there is still time for some good old ctrl+z.